requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scipy>=1.10.0
numpy>=1.24.0
Pillow>=10.0.0
transformers>=4.30.0

# Optional Dependencies (comment out if not needed)
face-recognition>=1.3.0  # For MyCropFace, MyFaceEmbedDistance
qwen-vl-utils>=0.0.1     # For MyQwenImageCaptioningNode, MyFolderFilterNode

# Segmentation models (optional)
# Uncomment if using MySegmentMask node
# scikit-image>=0.21.0
